♻️ Refactor the CMake file to follow more modern best practices
This includes generator expressions, correctly exports the targets for
use, allows the project to be used via `FetchContent`.

We also remove a bunch of the variable logic to just be configurable via
GNUInstallDirs (and let the defaults be used).

It is now possible to link via `SQLite3::Static` and `SQLite3::Dynamic`.
That said, it should probably be changed later to `SQLite3::Shared` to
follow common naming conventions.

Signed-off-by: Izzy Muerte <63051+bruxisma@users.noreply.github.com>